Unify inotify public interface across backends
